refactor(layout): two packages leave a name that was never theirs
`internal/platform` is the frozen host reference a release is qualified
against: Buildable, Reference, Facts, and the embedded platform.lock.json.
Two packages lived underneath it and had nothing to do with any of that.

`atomicfile` is a write primitive -- replace a file so no reader sees it
half-written -- used by the egress gateway and the capsule supervisor.
`githubactions` is the provider adapter. Neither is about a qualified
host, and neither is reachable from the package that is; a reader who
finds them under that path has been told a relationship that does not
exist, and the arch-test roster listed the atomic write beside the
vocabulary every layer shares while giving it a path implying otherwise.

The parent keeps its name because the name is right for what it holds:
"platform" is the vocabulary of linux/amd64, of the lock file, of
verify-platform and of the qualification record, and renaming it would
move a word that is load-bearing in five other places to fix a problem
that is not there. What was wrong was the two children inheriting it.

Import paths only. No identifier, behaviour, configuration or schema
change, and no CHANGELOG entry: nothing a deployment can observe moved.
