+++ internal/atomicfile/atomicfile.go
