+++ internal/command/doctor.go
@@ -12,7 +12,7 @@ import (
 	"github.com/rhobuild/runpool/internal/credential"
 	"github.com/rhobuild/runpool/internal/doctor"
 	"github.com/rhobuild/runpool/internal/engine/docker"
-	"github.com/rhobuild/runpool/internal/platform/githubactions"
+	"github.com/rhobuild/runpool/internal/githubactions"
 	"github.com/rhobuild/runpool/internal/store"
 )
